refactor: move Wallet components to a dir
