+++ renderer/containers/Wallet/Wallet.js
