+++ renderer/containers/Wallet/WalletStarter.js
