+++ renderer/containers/Wallet/index.js
@@ -1,0 +1,4 @@
+import Wallet from './Wallet'
+
+export default Wallet
+export WalletStarter from './WalletStarter'
